requirements.txt
# AI Photo Deduper
# 智能照片去重工具依赖

# 图像处理 (必需)
Pillow>=10.0.0

# 感知哈希算法 (核心功能)
imagehash>=4.3.1

# 可选: 进度条
tqdm>=4.65.0

# 可选: 更好的输出格式化
rich>=13.0.0
